Bookkeeping patch
This allows for bookkeeping. I'm keeping it with hardcoded aspects until it's finished. Basically I declared a new function and then called it from main. It should read and write to the proposed textfiles. Idk how to pick what to keep or not so we are gonna take this slow.